refactor: change behaviour for disabled state
Due to the way the disabled state is handled, a toggle will get the same color as unchecked as well as the cursor remaining to be a pointer. 

Too improve the UX I applied an opacity to the checkedBg color for better visibility and changed the icon to not-allowed as you would expect with a disabled element.